Revert "recordbench: fix signed-overflow UB (unsigned fields), not a compiler bug"
This reverts commit ca557d536cac497b171db29060de75a1b1e0ca0c.

+++ test/suites/recordbench/recordbench.c
@@ -12,21 +12,6 @@
  *
  * All field values are 15-bit-masked and the checksum 16-bit-masked, so it
  * matches on a 16-bit target and a 32-bit host.
- *
- * The fields MUST be `unsigned int`, not `int`.  Each masked field is in
- * [0,0x7fff], so an intermediate sum like `p->a + p->b` reaches up to 0xfffe.
- * With `int` fields on a 16-bit-int target (z80) that sum exceeds INT_MAX
- * (0x7fff) *before* the `& 0x7fff` mask is applied -> signed-overflow UB.
- * clang then legitimately emits `add nsw`, and InstCombine uses the nsw
- * assumption (a non-negative, non-overflowing sum is <= 0x7fff, so bit 15 is
- * already 0) to delete the now-"redundant" `& 0x7fff` mask.  The observed
- * symptom was exactly that: 2 of the 4 masks (on p->a and p->d) were dropped,
- * bit 15 leaked into the checksum, and the result came out 37157 (0x9125)
- * instead of 4389 (0x1125) -- a +0x8000 error.  This is NOT a compiler bug;
- * the "host-verified" constant was computed with 32-bit int where the same
- * sum (<= 0xfffe) never overflows.  `unsigned int` makes the wrap well
- * defined on both widths (no nsw -> masks preserved -> identical result),
- * which is what the "matches on 16-bit and 32-bit" contract above requires.
  */
 
 #include <stdio.h>
@@ -36,7 +21,7 @@
 #define ITERS  6000
 #define CHK    4389u        /* host-verified */
 
-struct st { unsigned int a, b, c, d, e; };
+struct st { int a, b, c, d, e; };
 static struct st S;
 
 static unsigned int churn(struct st *p, int it)
@@ -49,7 +34,7 @@ static unsigned int churn(struct st *p, int it)
         if (p->c & 1) p->d = (p->d + p->e) & 0x7fff;
         else          p->e = (p->e + p->a) & 0x7fff;
         p->b = (p->b + p->c) & 0x7fff;
-        chk = (chk + p->a + p->d) & 0xffffu;
+        chk = (chk + (unsigned int)p->a + (unsigned int)p->d) & 0xffffu;
     }
     return chk;
 }
